split out plotting functions into new file
